requirements.txt
fastapi
uvicorn
pillow
face_recognition
python-multipart